requirements.txt
langchain~=0.3.0
langchain-community~=0.3.0
langchain-core~=0.3.0
langchain_openai~=0.3.22
pydantic~=2.10.6
uvicorn~=0.34.3
fastapi~=0.115.12
langserve~=0.3.1
python-dotenv~=1.1.0